big change: changed the name of the packages and renamed a bunch of stuff
