+++ mobile/google-services.json
@@ -23,6 +23,58 @@
           }
         },
         {
+          "client_id": "1019793408847-pvfi6v6scnahbr4emnr7edbjfr2l770v.apps.googleusercontent.com",
+          "client_type": 1,
+          "android_info": {
+            "package_name": "com.communal_solutions.www.communalsolutions",
+            "certificate_hash": "bb4ba552918f080b9988c4ee5625d6e95067f609"
+          }
+        },
+        {
+          "client_id": "1019793408847-3qfb4g8nc38nu6dj72kki0l4k89tbain.apps.googleusercontent.com",
+          "client_type": 3
+        }
+      ],
+      "api_key": [
+        {
+          "current_key": "AIzaSyCL1G6Q3fQxNGjTtalx5aVFP5MLY1mhs94"
+        }
+      ],
+      "services": {
+        "analytics_service": {
+          "status": 1
+        },
+        "appinvite_service": {
+          "status": 2,
+          "other_platform_oauth_client": [
+            {
+              "client_id": "1019793408847-3qfb4g8nc38nu6dj72kki0l4k89tbain.apps.googleusercontent.com",
+              "client_type": 3
+            }
+          ]
+        },
+        "ads_service": {
+          "status": 2
+        }
+      }
+    },
+    {
+      "client_info": {
+        "mobilesdk_app_id": "1:1019793408847:android:4a3d9037d145a0ca",
+        "android_client_info": {
+          "package_name": "comcomsol.wixsite.communalsolutions1.communalsolutions"
+        }
+      },
+      "oauth_client": [
+        {
+          "client_id": "1019793408847-kcm0vbf20lu2iatb1ok5tjjte655kk6j.apps.googleusercontent.com",
+          "client_type": 1,
+          "android_info": {
+            "package_name": "comcomsol.wixsite.communalsolutions1.communalsolutions",
+            "certificate_hash": "95b2270591e26874041db96013dff389dac305e3"
+          }
+        },
+        {
           "client_id": "1019793408847-3qfb4g8nc38nu6dj72kki0l4k89tbain.apps.googleusercontent.com",
           "client_type": 3
         }
