--- mobile/src/androidTest/java/com/communal_solutions/www/communalsolutions/ExampleInstrumentedTest.kt
@@ -1,24 +1,0 @@
-package com.communal_solutions.www.communalsolutions
-
-import android.support.test.InstrumentationRegistry
-import android.support.test.runner.AndroidJUnit4
-
-import org.junit.Test
-import org.junit.runner.RunWith
-
-import org.junit.Assert.*
-
-/**
- * Instrumented test, which will execute on an Android device.
- *
- * See [testing documentation](http://d.android.com/tools/testing).
- */
-@RunWith(AndroidJUnit4::class)
-class ExampleInstrumentedTest {
-    @Test
-    fun useAppContext() {
-        // Context of the app under test.
-        val appContext = InstrumentationRegistry.getTargetContext()
-        assertEquals("com.communal_solutions.www.communalsolutions", appContext.packageName)
-    }
-}
