+++ mobile/src/debug/res/values/google_maps_api.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="google_maps_key" templateMergeStrategy="preserve" translatable="false">AIzaSyB8nRSQyUun8uNp5hd0yy4tRFlqsYcr3yM</string>
+    <string name="google_maps_key" templateMergeStrategy="preserve" translatable="false">AIzaSyAksyb9gAFJFMSjUMvnFd2ndsJGWVkv22k</string>
 </resources>
