--- mobile/src/main/java/com/communal_solutions/www/communalsolutions/Managers/MapsManager.kt
@@ -1,4 +1,0 @@
-package com.communal_solutions.www.communalsolutions.Managers
-
-class MapsManager {
-}
