+++ mobile/src/main/java/comcomsol/wixsite/communalsolutions1/communalsolutions/HelperFiles/Contact.kt
@@ -1,0 +1,15 @@
+package comcomsol.wixsite.communalsolutions1.communalsolutions.HelperFiles
+
+class Contact() {
+    var cName: String = "No Contact"
+    var cPhoneNum: String = "No Number"
+
+    constructor(name: String, phone: String) : this() {
+        cName = name
+        cPhoneNum = phone
+    }
+
+    override fun toString(): String {
+        return "Contact(cName='$cName', cPhoneNum='$cPhoneNum')"
+    }
+}
