+++ mobile/src/main/java/comcomsol/wixsite/communalsolutions1/communalsolutions/Managers/MapsManager.kt
@@ -1,0 +1,4 @@
+package comcomsol.wixsite.communalsolutions1.communalsolutions.Managers
+
+class MapsManager {
+}
