+++ releases/release/output.json
@@ -1,0 +1,1 @@
+[{"outputType":{"type":"APK"},"apkInfo":{"type":"MAIN","splits":[],"versionCode":1,"versionName":"1.0","enabled":true,"outputFile":"mobile-release.apk","fullName":"release","baseName":"release"},"path":"mobile-release.apk","properties":{}}]
