+++ wear/build.gradle
@@ -7,7 +7,7 @@ apply plugin: 'kotlin-android-extensions'
 android {
     compileSdkVersion 26
     defaultConfig {
-        applicationId "com.communal_solutions.www.communalsolutions"
+        applicationId "comcomsol.wixsite.communalsolutions1.communalsolutions"
         minSdkVersion 25
         targetSdkVersion 26
         versionCode 1
