--- wear/src/main/java/com/communal_solutions/www/communalsolutions/MainActivity.kt
@@ -1,15 +1,0 @@
-package com.communal_solutions.www.communalsolutions
-
-import android.os.Bundle
-import android.support.wearable.activity.WearableActivity
-
-class MainActivity : WearableActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-
-        // Enables Always-on
-        setAmbientEnabled()
-    }
-}
