+++ wear/src/main/java/comcomsol/wixsite/communalsolutions1/communalsolutions/MainActivity.kt
@@ -1,0 +1,15 @@
+package comcomsol.wixsite.communalsolutions1.communalsolutions
+
+import android.os.Bundle
+import android.support.wearable.activity.WearableActivity
+
+class MainActivity : WearableActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+
+        // Enables Always-on
+        setAmbientEnabled()
+    }
+}
